add shared types in types/ path - minor adjustments
