+++ src/routes/Images/ImagesView.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useSearchParams } from "react-router-dom";
 import { useQueryClient } from "@tanstack/react-query";
 import { useInfiniteRandomImages } from "./hooks";
-import type { CatImage } from "./api";
+import type { CatImage } from "../../types";
 import ImageModal from "./ImageModal";
 
 const ImagesView = () => {
